// src/app/auth/model/user.model.ts
export class User {
	email: string;
	password: string;
	passwordConf: string;
	primaryMobile: string;
	companyName: string;
	companyAddress: string;
	registrationtype: string;
	countryCode: string;
	name: string;
	username:string;
	_id: string;
	MCNumber: string;
	competition: string;
	message: string;
	prize: string;
	tc: string;
	date: Date;
	date2: Date;
	


}